chore: introduce trafficOfficer file
